requirements.txt
sqlalchemy
openai>=0.27,<1
python-dotenv
psycopg2-binary
